+++ examples/catalog/product/get.php
@@ -3,9 +3,9 @@
 require_once(dirname(__FILE__) . '/../../bootstrap.php');
 
 /**
-* preconditions:
-* - a default catalog exists
-*/
+ * preconditions:
+ * - a default catalog exists
+ */
 
 try {
     $products = $sdk->getCatalogService()->getProducts();