+++ examples/fixtures.php
@@ -1,13 +1,15 @@
 <?php
 
-use Shopgate\ConnectSdk\Dto\Base;
 use Shopgate\ConnectSdk\Dto\Catalog\Attribute;
 use Shopgate\ConnectSdk\Dto\Catalog\Attribute\Dto\Name;
 use Shopgate\ConnectSdk\Dto\Catalog\AttributeValue;
+use Shopgate\ConnectSdk\Dto\Catalog\Catalog;
+use Shopgate\ConnectSdk\Dto\Location\Location;
 use Shopgate\ConnectSdk\Dto\Catalog\Category;
 use Shopgate\ConnectSdk\Dto\Catalog\Inventory;
 use Shopgate\ConnectSdk\Dto\Catalog\Product;
 use Shopgate\ConnectSdk\Dto\Order\Order;
+use Shopgate\ConnectSdk\Exception\InvalidDataTypeException;
 
 const PARENT_CATALOG_CODE = 'BANA';
 const CATALOG_CODE = 'NARetail';
@@ -31,31 +33,32 @@ const INVENTORY_SKU_SECOND = 'Second Test';
 
 /**
  * @return array
+ *
+ * @throws InvalidDataTypeException
  */
 function provideParentCatalogs()
 {
     $parentCatalogs = [
-        'parentCatalogs' => [
-            new Base([
-                'code' => PARENT_CATALOG_CODE,
-                'name' => 'Team Banana Parent Catalog',
-                'isDefault' => true,
-                'defaultLocaleCode' => 'en-us',
-                'defaultCurrencyCode' => 'USD',
-            ])
-        ]
+        (new Shopgate\ConnectSdk\Dto\Catalog\ParentCatalog\Create())
+            ->setCode(PARENT_CATALOG_CODE)
+            ->setName('Team Banana Parent Catalog')
+            ->setIsDefault(true)
+            ->setDefaultCurrencyCode('USD')
+            ->setDefaultLocaleCode('en-us')
     ];
     return $parentCatalogs;
 }
 
 /**
  * @return array
+ *
+ * @throws InvalidDataTypeException
  */
 function provideCatalogs()
 {
     $catalogs = [
         'catalogs' => [
-            new Base([
+            new Catalog\Create([
                 'code' => CATALOG_CODE,
                 'parentCatalogCode' => PARENT_CATALOG_CODE,
                 'name' => 'North American Wholesale',
@@ -70,6 +73,8 @@ function provideCatalogs()
 
 /**
  * @return Category\Create[]
+ *
+ * @throws InvalidDataTypeException
  */
 function provideSampleCategories()
 {
@@ -102,6 +107,8 @@ function provideSampleCategories()
 
 /**
  * @return Attribute\Create[]
+ *
+ * @throws InvalidDataTypeException
  */
 function provideSampleAttributes()
 {
@@ -156,6 +163,8 @@ function provideSampleAttributes()
 
 /**
  * @return Inventory\Create[]
+ *
+ * @throws InvalidDataTypeException
  */
 function provideSampleInventories()
 {
@@ -222,6 +231,8 @@ function provideSampleDeleteInventories()
 
 /**
  * @return Product\Create[]
+ *
+ * @throws InvalidDataTypeException
  */
 function provideSampleProducts()
 {
@@ -248,11 +259,13 @@ function provideSampleProducts()
 
 /**
  * @return array
+ *
+ * @throws InvalidDataTypeException
  */
 function provideLocations()
 {
     return [
-        new Base([
+        new Location\Create([
             'code' => LOCATION_CODE,
             'name' => 'Warehouse 1',
             'status' => 'active',
@@ -265,6 +278,11 @@ function provideLocations()
     ];
 }
 
+/**
+ * @return array
+ *
+ * @throws InvalidDataTypeException
+ */
 function provideSampleOrders()
 {
     $order = new Order\Create([