+++ src/Dto/Catalog/AttributeValue/Get.php
@@ -27,11 +27,6 @@ use Shopgate\ConnectSdk\Dto\Catalog\AttributeValue;
 /**
  * Dto for attribute values response.
  *
- * @method string getCode()
- * @method int getSequenceId()
- * @method Dto\Name getName()
- * @method Dto\Swatch getSwatch()
- *
  * @codeCoverageIgnore
  */
 class Get extends AttributeValue
@@ -42,7 +37,7 @@ class Get extends AttributeValue
     protected $schema = [
         'type'                 => 'object',
         'properties'           => [
-            'swatch' => ['$ref' => Dto\Swatch::class]
+            'swatch' => ['$ref' => Dto\Swatch::class, 'skipValidation' => true]
         ],
         'additionalProperties' => true,
     ];