+++ tests/Integration/Dto/Catalog/InventoryTest.php
@@ -26,9 +26,9 @@ use Shopgate\ConnectSdk\Dto\Catalog\Inventory;
 use Shopgate\ConnectSdk\Exception\NotFoundException;
 use Shopgate\ConnectSdk\Exception\RequestException;
 use Shopgate\ConnectSdk\Exception\Exception;
-use Shopgate\ConnectSdk\Tests\Integration\CatalogTest;
+use Shopgate\ConnectSdk\Tests\Integration\AbstractCatalogTest;
 
-class InventoryTest extends CatalogTest
+class InventoryTest extends AbstractCatalogTest
 {
     /**
      * @throws Exception