+++ tests/Unit/Dto/Order/Order/GetTest.php
@@ -83,6 +83,7 @@ class GetTest extends TestCase
             'total' => 105,
             'date' => 'today',
             'submitDate' => 'today',
+            'updateDate' => 'today',
             'sourceDevice' => 'mobile',
             'sourceIp' => '111.111.111.111',
             'fulfillmentGroups' => [
@@ -109,7 +110,7 @@ class GetTest extends TestCase
                                     'tracking' => 'test-tracking-code-two',
                                     'pickUpBy' => 'Johnny Bravo',
                                     'labelUrl' => 'test-label-url',
-                                    'fulfilledDate' => 'tomorrow',
+                                    'fulfillmentDate' => 'tomorrow',
                                     'packageItems' => [
                                         [
                                             'id' => 'product-id-one',
@@ -135,7 +136,14 @@ class GetTest extends TestCase
                         'name' => 'Product One',
                         'image' => 'image-url',
                         'price' => 100,
-                        'currencyCode' => 'USD'
+                        'currencyCode' => 'USD',
+                        'options' => [
+                            [
+                                'code' => '146',
+                                'name' => 'Color',
+                                'value' => 'Red',
+                            ]
+                        ]
                     ],
                     'currencyCode' => 'USD',
                     'shippingAmount' => 5,
@@ -209,6 +217,7 @@ class GetTest extends TestCase
         $this->assertEquals($entry['total'], $get->getTotal());
         $this->assertEquals($entry['date'], $get->getDate());
         $this->assertEquals($entry['submitDate'], $get->getSubmitDate());
+        $this->assertEquals($entry['updateDate'], $get->getUpdateDate());
         $this->assertEquals($entry['sourceDevice'], $get->getSourceDevice());
         $this->assertEquals($entry['sourceIp'], $get->getSourceIp());
         $actualFulfillmentGroup = $get->getFulfillmentGroups()[0];
@@ -254,8 +263,8 @@ class GetTest extends TestCase
         $this->assertEquals($expectedFulfillmentPackage['pickUpBy'], $actualFulfillmentPackage->getPickUpBy());
         $this->assertEquals($expectedFulfillmentPackage['labelUrl'], $actualFulfillmentPackage->getLabelUrl());
         $this->assertEquals(
-            $expectedFulfillmentPackage['fulfilledDate'],
-            $actualFulfillmentPackage->getFulfilledDate()
+            $expectedFulfillmentPackage['fulfillmentDate'],
+            $actualFulfillmentPackage->getFulfillmentDate()
         );
         $actualPackageItems = $actualFulfillmentPackage->getPackageItems()[0];
         $expectedPackageItems = $expectedFulfillmentPackage['packageItems'][0];
@@ -290,15 +299,22 @@ class GetTest extends TestCase
         $this->assertEquals($expectedLineItem['overrideAmount'], $actualLineItem->getOverrideAmount());
         $this->assertEquals($expectedLineItem['extendedPrice'], $actualLineItem->getExtendedPrice());
         $this->assertEquals($expectedLineItem['price'], $actualLineItem->getPrice());
+
+        $actualProduct = $actualLineItem->getProduct();
         $this->assertInstanceOf(OrderDto\LineItem\Product::class, $actualLineItem->getProduct());
-        $this->assertEquals($expectedLineItem['product']['code'], $actualLineItem->getProduct()->getCode());
-        $this->assertEquals($expectedLineItem['product']['name'], $actualLineItem->getProduct()->getName());
-        $this->assertEquals($expectedLineItem['product']['image'], $actualLineItem->getProduct()->getImage());
-        $this->assertEquals($expectedLineItem['product']['price'], $actualLineItem->getProduct()->getPrice());
-        $this->assertEquals(
-            $expectedLineItem['product']['currencyCode'],
-            $actualLineItem->getProduct()->getCurrencyCode()
-        );
+        $this->assertEquals($expectedLineItem['product']['code'], $actualProduct->getCode());
+        $this->assertEquals($expectedLineItem['product']['name'], $actualProduct->getName());
+        $this->assertEquals($expectedLineItem['product']['image'], $actualProduct->getImage());
+        $this->assertEquals($expectedLineItem['product']['price'], $actualProduct->getPrice());
+        $this->assertEquals($expectedLineItem['product']['currencyCode'], $actualProduct->getCurrencyCode());
+
+        $productOptions = $actualProduct->getOptions();
+        $this->assertTrue(is_array($productOptions));
+        $productOption = $productOptions[0];
+        $this->assertEquals($expectedLineItem['product']['options'][0]['code'], $productOption->getCode());
+        $this->assertEquals($expectedLineItem['product']['options'][0]['name'], $productOption->getName());
+        $this->assertEquals($expectedLineItem['product']['options'][0]['value'], $productOption->getValue());
+
         $actualHistoryItem = $get->getHistory()[0];
         $expectedHistoryItem = $entry['history'][0];
         $this->assertInstanceOf(OrderDto\HistoryItem::class, $actualHistoryItem);