+++ tools/fixtures/schema.sql
@@ -7,6 +7,7 @@ DROP DATABASE IF EXISTS location;
 DROP DATABASE IF EXISTS merchant;
 DROP DATABASE IF EXISTS customer;
 DROP DATABASE IF EXISTS import;
+DROP DATABASE IF EXISTS webhook;
 DROP DATABASE IF EXISTS omnichannel_order;
 DROP DATABASE IF EXISTS omnichannel;
 DROP DATABASE IF EXISTS omnichannel_auth;
@@ -18,6 +19,7 @@ CREATE DATABASE omnichannel;
 CREATE DATABASE omnichannel_auth;
 CREATE DATABASE import;
 CREATE DATABASE omnichannel_order;
+CREATE DATABASE webhook;
 
 DROP TABLE IF EXISTS location.`Location`;
 
@@ -618,3 +620,23 @@ CREATE TABLE omnichannel_user.`UserMerchant`
 ) ENGINE = InnoDB
   DEFAULT CHARSET = utf8mb4;
 
+DROP TABLE IF EXISTS omnichannel_user.`UserMetric`;
+
+CREATE TABLE omnichannel_user.`UserMetric`
+(
+    `UserMetricID` bigint(20) unsigned NOT NULL AUTO_INCREMENT,
+    `UserID`       char(36)            NOT NULL,
+    `MerchantID`   char(36)                     DEFAULT NULL,
+    `Key`          varchar(255)        NOT NULL DEFAULT '',
+    `Value`        varchar(255)                 DEFAULT NULL,
+    `CreateBy`     varchar(255)        NOT NULL DEFAULT '',
+    `CreateDate`   datetime            NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    `UpdateBy`     varchar(255)                 DEFAULT NULL,
+    `UpdateDate`   datetime                     DEFAULT NULL ON UPDATE CURRENT_TIMESTAMP,
+    `DeleteBy`     varchar(255)                 DEFAULT NULL,
+    `DeleteDate`   datetime                     DEFAULT NULL,
+    PRIMARY KEY (`UserMetricID`),
+    UNIQUE KEY `UserMetric_Key` (`UserID`, `MerchantID`, `Key`),
+    CONSTRAINT `UserMetric_UserID` FOREIGN KEY (`UserID`) REFERENCES `User` (`UserID`) ON DELETE CASCADE ON UPDATE CASCADE
+) ENGINE = InnoDB
+  DEFAULT CHARSET = utf8mb4;