+++ tools/pubsubfiller.php
@@ -41,6 +41,7 @@ $topics = [
         'webhookTransInventoryReservationSettled-development',
     ],
     'salesOrderAdded-development' => [
+        'orderSvcSalesOrderAdded-development',
         'routingSvcSalesOrderAdded-development',
         'webhookSalesOrderAdded-development',
         'webhookTransSalesOrderAdded-development',
@@ -83,6 +84,9 @@ $topics = [
     'webhookRouted-development' => [
         'webhookTransWebhookRouted-development',
     ],
+    'orderNotPickedUp-development' => [
+        'webhookTransOrderNotPickedUp-development',
+    ]
 ];
 
 foreach ($topics as $topic => $subscriptions) {